requirements.txt
twitchio
websocket-client
python-dotenv
requests
sentence-transformers


langchain
langchain-core
langchain-huggingface
langchain-community